examples/problem_grader/requirements.txt
Django==1.5
mimeparse==0.1.3
lxml==3.0.1
pyyaml==3.10
uuid==1.30
south==0.7.6
path.py==3.0
requests==1.1.0
django-pipeline==1.3.6
nose==1.2.1
django-nose==1.1
simplejson==3.1.3
slumber==0.6.0